Rewrite custom response classes as wrappers
This rewrite net-simplifies the combination of responses and various
HTTP methods. The overall gist is that base client methods always
return GlobusHTTPResponse objects, and the ability to pass
`response_class` is removed. In order to still support advanced
response types (ActivationRequirementsResponse, OAuthTokenResponse,
etc), these types become proxy or wrapper objects which contain a
GlobusHTTPResponse.

To make defining these simpler, a new class, GlobusHTTPResponseProxy
is added, from which these inherit. The proxy just pulls attributes
off of the inner response and re-exposes them. Both GlobusHTTPResponse
and GlobusHTTPResponseProxy inherit from GlobusResponse, which lets
them share implementations of `__getitem__`, `__contains__`, and
`get()` with respect to their `data` property.
GlobusResponse becomes abstract, with `data` as an abstract property.

As a result, where we would previously provide a desirable response
class like so:

    self.get("/foo", response_class=FooResponse)

The same idea is expressed in a more "ordinary" and idiomatic way like
so:

    FooResponse(self.get("/foo"))

The following included changes are related to this:

- PaginatedResponse no longer needs to take a `response_class` to pass
  through -- it just wraps data in IterableTransferResponse

- The once-optional `client` argument to GlobusHTTPResponse is now
  required, since all of this was being refactored anyway

- Type overloads for the BaseClient methods are all removed, because
  they now are guaranteed to always return GlobusHTTPResponse

- AuthClient.oauth2_token needs to allow the use of variable wrapper
  types to support dependent token responses while still returning a
  default of an OAuthTokenResponse. It therefore implements and has
  type overloads for `response_class`, similar to what was previously
  done in BaseClient