Enhance and improve error parsing
GlobusAPIError gains a new property, `info`, which contains variables
for `authorization_parameters` and `consent_required`, each of which
points to an "ErrorInfo" object of the appropriate type.

This encapsulates any parsing of the response data (beyond basic JSON
parsing) in a dedicated set of classes.

Because the `.info` attribute is added to all errors, the information
about the type of error (e.g. "consent required") is independent from
the particular API error class.
Effectively, this is composition ("has a") rather than inheritance
("is a").

Additionally, the "base" error parsing logic can be simplified. Start
by assuming that `code` and `message` are, as for text responses,
"Error" and <response>.text . Then overwrite them when parsing JSON
*if* we find relevant properties.
This reduces the risk that parsing fails while handling an API
response with some -- but not all -- of our expected properties. All
fields are populated, and it's possible to pass `details`, `message`,
and `code` independently. (e.g. an error with `message` but no `code`
will be handled gracefully)