Remove unreachable check from client __init__
There's a check for `service_name or base_url` being truthy. However,
this will always be truthy with the current state of the code.
`base_url` is optional, but `service_name` is a string with a special
default of `"_base"`.

The only way to trigger this check (with code which passes type
checking) is to do something very intentionally wrong:

    class MyClient(BaseClient):
        service_name = ""