Update to use the latest 'responses' on py3.7+
This is an update to the testing requirements to use different
versions of `responses` on different python versions. It *should*
allow use of the latest `responses` version with `globus_sdk._testing`
without warnings or errors.

As a related tweak, this change revealed that we still have one test
which was using `six` (installed via the `responses` dependency). It
is easily removed.