Add some tests for non-default RequestsMock
This is part of _testing being tested, so it's hard to see a way of
doing it without getting tied to some of the methods we have
throughout the SDK. Choose a relatively simple one (get_group) as an
example.

Not all methods are tested here, just a couple of essentials.