Improve TransferData and DeleteData null handling
Convert various argument usages to the null-removing setters and
ensure that `add_item` removes null arguments instead of passing them
to the API.

Allow transfer_client=None in order to support a usage in which
`submission_id` is omitted for both TransferData and DeleteData.

Convert `skip_activation_check` to an Optional with a default of
None so that it will be omitted by default.