TransferData and DeleteData allow kwarg-only usage
In these helper objects, set `transfer_client`,
`source_endpoint`, `destination_endpoint`, and `endpoint` to optional
with defaults of `None`. They are now optional positional arguments,
meaning that they can be passed as named parameters, instead of
requiring a first positional argument of `None` for the
null-TransferClient usage.

As a result, the `source_endpoint`, `destination_endpoint`, and
`endpoint` parameters need to be explicitly checked at runtime to
ensure they are not passed as `None`. Although this slightly weakens
the capabilities of type checking, it allows for better runtime
semantics and usage in the good case.

The testsuite is updated to test these behaviors and ensure that bad
arguments are rejected. As the usage is now simpler, some testing
helpers are no longer useful and have been removed.

Two related but distinct changes are added here in support of Timer
usage:
- TimerJob now logs an info message on init (useful for
  troubleshooting)
- `recursive_symlinks` is now Optional with a default of `None`,
  meaning it will be omitted from TransferData by default