Add a minimal FlowsClient supporting list_flows
FlowsClient is defined with the attributes and configuration needed
to talk to the Globus Flows API. A single paginated method named
"list_flows" is supplied, with only some of its (many) parameters
provided as keyword arguments.

A data stub in _testing and a single test for the "simple listing
case" are included the exercise the code and ensure the resolution of
the service URL and other plumbing works as expected.

_generate_init is updated with the new module and its attributes.

One notable quirk of this client: the method is paginated, but the
pagination parameter is named "pagination_token" rather than the more
typical "marker". For this reason, the parameter name is now
configurable when instantiating a MarkerPaginator.
The paginated usage is not tested as of yet.