Support python 3.11
In the changelog, this is called out as "official support" since the
code all worked on 3.11 prior to this change.

Places touched:
- CI testing job on linux (all pythons)
- CI testing on windows (latest python only)
- CI testing on macos (latest python only)
- readthedocs config
- pypi classifiers
- tox testenvs

CI builds should be examined to see what might fail in particular.
Local testing under pyenv+tox works, but it's always possible that
something will be different in the CI environment.