Add a close() method to SQLiteAdapter
This is being done to fix the behavior of tests which are leaking
resources. The leaked connections can cause tests to fail because
there's still a handle on the `.db` file when the temp dir tree is
being removed.

The method makes it possible to close the connection through a public
interface. Rather than adding a context manager interface, for now the
testsuite operates by having a fixture which auto-closes SQLiteAdapter
instances during test teardown.