Convert GCS get_collection_list tests to new style
This change moves the tests for this method into a dedicated module,
and moves the test data from the older `fixture_data/` dir to the
`globus_sdk._testing.data` subpackage.

The oddity here is that get_collection_list is being used to test
error parsing for a GCS 403. It's not immediately clear where this
should go, so it's bundled with these other tests for now. (Perhaps
this test is good but should be augmented by additional unit tests.)