Improve str(OAuthTokenResponse)
This makes the str() of an oauth response object more verbose,
clarifying that `by_resource_server` is a field (not the set of
top-level keys) and including mention of `id_token`. The output is
also now more YAML-like, and not just an undecorated JSON dump.

In practice, the `id_token` can be very long, so this implementation
truncates it to the first 10 chars and notes it as `... (truncated)`.

No other fields are added to the stringifier as of yet.

resolves #639