Add scope parsing to Scope object
The core of this change is that `MutableScope` has been renamed to
`Scope` and now has a pair of parsing functions as classmethods:
`Scope.parse` and `Scope.deserialize`. The first parses any valid
scope string to request, meaning it may be a space delimited list of
scopes, and the second parses any valid scope string with a single
top-level scope.

As an inverse of `deserialize()`, the stringifier has been renamed to
`serialize()`, and `__str__()` now just calls `serialize()`.

The parsing itself requires several new internal components. A
tokenizer handles converting strings into lists of tokens, e.g.
converting `foo[bar *baz]` into
`["foo", "[", "bar", "*", "baz", "]"]`.
The tokens are small objects pairing strings with a type, which makes
it simpler for the parser to handle various cases. The parser is
minimally stateful, has no recursion or backtracking, operates with
limited lookahead, and is generally meant to be the smallest and
simplest thing which can do the job that it does.

The parse result, a `Scope`, can now represent the true tree-like
structure of scopes. `Scope.dependencies` is a list of child nodes in
the tree, a list of  `Scope`s. This has implicitations for
`Scope.add_dependency` in that a `Scope` is now a valid input. As a
result of this introduction, the use of
`add_dependency("foo", optional=True)` is considered deprecated and
now emits a DeprecationWarning. Building dependencies in the current
version should be done either with (newly parseable) strings or with
`Scope` objects. i.e. The aforementioned usage can be replaced with
either `add_dependency("*foo")` or
`add_dependency(Scope("foo", optional=True))`.

Some related changes included here:
- `globus_sdk.scopes` is now a subpackage (it was getting a bit long)
- the internal type alias for ScopeCollectionType is now kept in
  `_types` for simpler universal access within the SDK
- many more unit tests are added for the scope parsing and
  serialization functionality
- documentation is updated to refer to these as "Scope objects" rather
  than "MutableScopes"
- various doc fixes, including missing scope data in the scopes docs
  and a bug in the listknownscopes custom directive