Add `Scope.__contains__` with tests
`Scope(...) in Scope(...)` is meant to indicate "permissions
containment", meaning that the permissions requested by the first
scope are covered by the permissions requested by the second scope.

That notion of containment leads to the following defined behaviors:
- the top-level scope strings must match
- either both have matching optional=t/false OR only the
  "contained" scope is optional
- each dependency needs to be contained in the "containing" scope's
  dependencies
- it doesn't make sense to ask if a non-Scope value is in a Scope

This definition feeds directly into an implementation and a large
suite of tests to check correctness.