First version of LocalGlobusConnectServer object
This adds a start on a local GCS object which looks at the canonical
`info.json` file and pulls attributes out of it in a type-safe way.

Primarily in order to support testing, the path to the `info.json`
file can be set as an argument to the LocalGlobusConnectServer
initializer.