Make Scope "contains" a private helper method
We retain all of the tests, but the method is renamed to `_contains`,
removed from public docs, and commented as a "non-authoritative
convenience function". In particular, after discussion we do not
believe that it is valuable to try to build fully accurate
client-side scope comparison when this cannot be authoratative with
respect to a user's current consents in Auth.

For now, the method will exist and it will be available for our own
use-cases (e.g. in globus-cli), but not for general use.

Also, improve documentation of `add_dependency` to avoid (confusing)
chaining usage.