Use a sorted list for lazy import test params
This fixes an issue on py3.11 in which the order of this frozenset is
no longer fixed (it was never guaranteed) and the result is that a
distributed test run delivers test parameters in different orders.
pytest-xdist fails if workers see different parameters -- even if it
is only order that varies.