Parse 'session_required_policies' to a list
Rather than keeping the string which an API may pass back, convert it
to the list of IDs by comma-splitting the value. If the value is not a
string, the comma split won't be possible, so this is handled by
treating anything other than a string as invalid (and setting the
attribute to None).

This also makes this property a little bit safer than many of our
other computed properties, which are cast to the "right" types by the
SDK but aren't validated.

The change is technically backwards incompatible, but surrounds the
use of a feature which is not in production use yet, and we have
therefore agreed upon pragmatic treatment of this case as not
requiring a major version bump.

The test changes include updates to check the parsed value(s), but
also a general update to the ErrorInfo tests to split a monolithic
test into more individual cases.