Update annotations to "future" style via pyupgrade
This was always going to be a multi-step process to add `__future__`
imports, rewrite, and iterate. However, there's an additional
wrinkle: unfortunately, `pyupgrade` does not support rewrites on
`import typing as t`-style annotations.
Therefore, the following procedure includes rewrites away from and
back to this style.

1. Set future annotations in all files

Run a script to inject `from __future__ import annotations` in all
non-`__init__.py` python files in the src tree.

2. Convert `import typing as t` style to `import typing` style

This was done with a combination of scripts and macros.

3. Run `pre-commit run -a` to converge

Converge, converge, converge. `black`, `isort`, `pyupgrade`, until
they get along.

4. Convert `import typing` style back to `import typing as t` style

Again, scripts and editor macros. Converge again, as the rewrites may
fold out and fold in with `black`, especially.

5. Update _generate_init.py to not check python version

The TYPE_CHECKING branch was combined with the py3.6 branch. Remove
that logic for better clarity.