Enable use of MutableScopes in client methods
The use of a MutableScope object, iterable thereof, or mixed iterable
of MutableScope and str objects, was already supported by the
underlying OAuth2 Flow Manager objects. However, the types used there
were not "propagated upwards" to the auth client methods which wrap
those manager objects.

Firstly, this change updates the annotations for the passthrough
methods and adds a typing test which ensures that the correct types
are allowed and defined by the ScopeCollectionType helper.

Secondarily, this adds the same capability to the
``oauth2_client_credentials_tokens`` method to make our scope
interfaces more uniform. Unlike the other cases, this isn't just a
typing change -- a new runtime behavior is being introduced here.
As a result, a new test is wanted to validate that behavior. It turns
out that this new test is the first test for this method, so fixture
data must be included.