Switch to GHA automated package publishing
We run a workflow to publish to test-pypi on tag, and a workflow to
publish to pypi on GitHub release.

The tox environment for publishing has been removed (it can always be
restored if it turns out to be necessary in the future).

`make release` has been replaced with `make tag-release`, which is
slightly more clear about where this step fits into the workflow, and
no longer tries to run the publishing step.

The `RELEASING.md` doc has been updated to accord with these details.