Simplify v4 deprecation warnings and fix tests
Converting the deprecation warning behavior to always-on using
DeprecationWarning as the base class caused some tests (which trigger
the relevant warning) to fail. In these cases, exercise both runtime
behaviors with and without triggering the warning, but mark the
warning variant with a `pytest.warns` assertion.

In order to support some of the fixed tests, a new fixture is added in
`globus_sdk._testing`.