Cleanup Local GCP tests and add data to fixtures
GlobusConnectPersonalOwnerInfo tests relied on the presence of
response payloads from get_identities with a few characteristics not
otherwise seen in the response data. One is a response for a
globusid.org user, and the other is an empty response. These two are
added to the test fixtures, and tests are converted to their use.