Improve SearchClient docs, deprecate entry methods
The create_entry and update_entry methods are misleadingly named. We
may or may not be able to remove these aliasing APIs from the service,
but we should at least discourage their use via the SDK. Redirect
users of these methods to Ingest, which is the only behavior offered
by all three of these APIs.

Even if we were to support this operation in the SDK, having a single
API with two methods is confusing because it suggests that there is
some difference.

Instead, we would probably want to implement `write_entry` if we need
a succcessor method.