Add 'refresh_tokens' to dependent token method
Dependent token grants support `access_type=offline`, which is
represented elsewhere in the SDK as `refresh_tokens=True`. Add support
for this parameter/translation as it currently exists. The default
(False) will be translated to parameter omission rather than
`access_type=online`.

In addition to the parameter change, test data and two tests are added
to test the method with and without this option.