Begin using tabs to refactor method docs
This is a combined set of several minor changes which support a new
way of writing docs for client methods. The overall goal is to make
use of sphinx-inline-tabs to present several pieces of information in
a way which is compact by default, expandable, and allows us to put
some more structure to what is currently a long scroll of text.

To wit:
- include the sphinx-inline-tabs extension for sphinx
- cleanup our custom sphinx directives to use docutils' built-in
  option parsing
- add new custom directives for several small functions:
  - expandtestfixture: expands a JSON or text body from a test fixture
    as a code-block, for use as example data
  - extdoclink: link to an external doc location (only supports
    docs.globus.org right now)
  - paginatedusage: emit the standard paginated usage info text
- select two methods which are updated to use the new documentation
  capabilities: TransferClient.get_submission_id and
  SearchClient.search. These demonstrate the behavior and what is now
  possible

It should be noted that `get_submission_id` has no usage example while
`search` does, that `search` is paginated while `get_submission_id` is
not, and that both refer to docs site docs and have relevant test
fixture data for examples. Therefore, they demonstrate the flexibility
of this setup to handle different kinds of methods, but also share
some commonalities.

The original drive of this work was `expandtestfixture`, but once that
was achieved, it demands some solution for better presentation, like
tabs. Once tabs are in place, they conflict with the default behavior
for the pagination decorator (which appends to the docstring), so some
accommodation is needed. A similar need precipitates the refactor of
the external doc decorator into a sphinx directive.