Convert NativeAppAuthClient to tab doc style
doc_api_method was only used for oauth2_start_flow, but in this place
we now use the tabs for documentation and the API Info section is
populated with specific and appropriate content.

doc_api_method supported passing in a format string, a behavior which
does not seem fully necessary and is intentionally omitted from
extdoclink.