Fix extdoclink to generate anonymous links
Links ending with `_` are made into reference targets, whereas those
ending with `__` are not. This "anonymous reference" behavior is part
of the RST specification, not only a sphinx behavior.

When sphinx sees multiple targets with the same name in a document, it
emits a warning, which we turn into an error with `-W`.

The duplicates easily occur with these links, which often match method
names. Using the `__` behavior dodges the problem.