Fix the return type annotation for get_identities
This was annotated as GlobusHTTPResponse when it should have been
GetIdentitiesResponse. The result is that type checkers could not see
that an iterable type was returned.

A new mypy test case confirms that the type is now seen correctly, as
well as validating several arguments and usages for get_identities.