Make HTTP responses and errors more similar
Ensure that these provide the same suite of properties for interacting
with HTTP response data. Importantly, the error class gains `text` (to
replace `raw_text`) and `content_type`.

Furthermore, accessing `raw_text` will now emit a RemovedInV4Warning
flavor of DeprecationWarning, indicating that we intend to remove the
property in SDK v4.

We could take the similarity further, making the error objects use
`data` to replace `raw_json`, but it is not clear that the "data part"
of the responses should be made uniform.

To encapsulate the suite of matching properties, the type helpers
module now includes a protocol class, `ResponseLike`. `ResponseLike`
is not exported by the SDK at the top level but is used to implement
testing to confirm that the types match. (This is open to future
change.)

Additionally, ResponseLike defines, and both classes are updated to
provide, `binary_content`. This points to `requests.Response.content`,
which is the response body as a bytestring.

Testing-wise, several tests are modified or introduced here:
- tests which access `raw_text` on errors expect RemovedInV4Warning
- tests now check `binary_content` for responses and errors
- mypy test cases enforce that GlobusAPIError and GlobusHTTPResponse
  both implement ResponseLike