Convert project metadata to declarative setup.cfg
setuptools supports declarative metadata in setup.cfg, replacing the
imperative style in setup.py. This results in a net-reduction in
complexity for us, as we no longer have to maintain code for loading
the readme or version data into the setuptools context.

Some specialized setuptools directives support pulling the version
number from a package attribute and reading the description from the
readme.

The rest of the metadata is a port of our setup.py content.

For broader compatibility with installations which explicitly invoke
`setup.py`, a stub setup.py is still included, as is typical for
packages which want to minimize the impact of a move to declarative
metadata.

Two semantic changes are made here, on the premise that neither will
be impactful for SDK users:
1. The vestigial 'dev' extra is removed. It contained no dependencies
   and there is no longer cause for anyone to install
   'globus-sdk[dev]', which was purely for dev/testing purposes
2. The MANIFEST.in has been removed, which results in our tox.ini not
   being included in the built package distributions.