Make authorization_parameters more type-safe
When parsing authorization_parameters, avoid previous assignments
which assumed that values had the correct types. Instead, always check
values against the expected type(s) and convert them to None with a
warning log if they are malformed.

In service of this improvement in strictness, the class has been
refactored to have a more verbose but highly unambiguous structure.