Add `construct_error` to `_testing` subpackage
This helper replaces several "make_response" fixtures in error
testing, not only making things more uniform and explicit within the
SDK tests, but also making it possible for `_testing` users to
leverage this art.

`construct_error` replaces an entire error-testing `conftest.py` file
and is generally a simplification. Several tests become more verbose
as they now contain more of the test data.

The type of `construct_error` needs to be declared with overrides to
allow for the `error_class` to indicate the return type.